Book Store sim 4.0

+++ src/BookShopSim.java
@@ -12,7 +12,7 @@ public class BookShopSim {
 
         Scanner sc = new Scanner(System.in);
         System.out.print("Enter balance: ");
-        double balance = sc.nextInt();
+        double balance = sc.nextDouble();
         printAll(code, name);
         findAndBuyBook(name, quantity, price, balance);
         System.out.print("Thank you for visiting the shop!");
